requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
crispy-tailwind==0.5.0
django==4.1.4
django-crispy-forms==1.14.0
django-extra-views==0.14.0
django-widget-tweaks==1.4.12
sqlparse==0.4.3 ; python_version >= '3.5'
